backend/requirements.txt
# Django and core packages
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-gis==1.0

# PostgreSQL and PostGIS
# psycopg2-binary==2.9.9

# CORS headers
django-cors-headers==4.3.1

# WebSocket support
channels==4.0.0
channels-redis==4.1.0

# LiDAR processing
laspy==2.5.1
PDAL==3.2.3

# Geospatial libraries
# GDAL==3.7.3
# Fiona==1.9.5
Shapely==2.0.2
pyproj==3.6.1

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.2
numpy==1.26.2
pandas==2.1.3
joblib==1.3.2

# Data processing
scipy==1.11.4

# Cloud storage (optional)
boto3==1.29.7  # For AWS S3
google-cloud-storage==2.13.0  # For GCS

# Utilities
python-dotenv==1.0.0
pillow==10.1.0
celery==5.3.4  # For async task processing
redis==5.0.1

# Development tools
ipython==8.18.1
django-debug-toolbar==4.2.0

# API documentation
drf-yasg==1.21.7
